+++ src/main/java/org/jd/core/v1/service/fragmenter/javasyntaxtojavafragment/util/StringUtil.java
@@ -49,8 +49,8 @@ public class StringUtil {
                         default:
                             if (c < ' ') {
                                 sb.append("\\0");
-                                sb.append((char)('0' + ((int)c >> 3)));
-                                sb.append((char)('0' + ((int)c & 7)));
+                                sb.append((char)('0' + (c >> 3)));
+                                sb.append((char)('0' + (c & 7)));
                             } else {
                                 sb.append(c);
                             }
