+++ src/test/java/org/jd/core/v1/CfrTest.java
@@ -48,15 +48,15 @@ public class CfrTest extends TestCase {
             }
 
             void test1() {
-                long b = (long) (double) getLong();
+                long b = (long) getLong();
                 System.out.println(b == getLong()); // Prints "false"
             }
             void test2() {
-                long b = (long) (double) l;
+                long b = (long) l;
                 System.out.println(b == l); // Prints "false"
             }
             void test3() {
-                long b = (long) (double) L;
+                long b = (long) L;
                 System.out.println(b == L); // Prints "false"
             }
         }
